mkinitrz: Rework adding libraries
The previous implementation read the elf section with all the
shared libraries and then used dl::open on each to follow the
linked chain. The problem with this approach is that
instead of getting the name of the linked library (i.e.
the symlink libcrypto.so.3) we got the full path
(/lib64/libcrypto.so.3.1.3 on openSUSE), skipping the symlink.

Instead, only use the library name and manually search in the
hardcoded library paths.